+++ packages/connext/src/Connext.ts
@@ -11,13 +11,13 @@ import {
   type TransactionFilter,
   compressJson,
 } from '@rabbitholegg/questdk'
-import { type Address } from 'viem'
+import { type Address, zeroAddress } from 'viem'
 import { XCALL_ABI_FRAGMENTS } from './abi.js'
 import { ConnextContract } from './contract-addresses.js'
 
 let _chainDataCache: Map<string, ChainData> | null = null
 
-const ETH_TOKEN_ADDRESS = '0x0000000000000000000000000000000000000000'
+const ETH_TOKEN_ADDRESS = zeroAddress
 
 const _getChainData = async () => {
   if (!_chainDataCache) {
@@ -28,82 +28,56 @@ const _getChainData = async () => {
   return _chainDataCache
 }
 
-export const getWETHAddress = async (chainId: number) => {
-  const chains = await _getChainData()
-  const domainId = chainIdToDomain(chainId)
-  const chainData = chains?.get(String(domainId))
-  const assets = Object.keys(chainData?.assetId || {})
-
-  let wethAddress
-  for (const address of assets) {
-    if (chainData?.assetId[address].symbol === 'WETH') {
-      wethAddress = address
-      break
-    }
-  }
-  return wethAddress
-}
-
 export const bridge = async (
   bridge: BridgeActionParams,
 ): Promise<TransactionFilter> => {
-  const {
-    sourceChainId,
-    destinationChainId,
-    contractAddress,
-    tokenAddress,
-    amount,
-    recipient,
-  } = bridge
+  const { sourceChainId, destinationChainId, tokenAddress, amount, recipient } =
+    bridge
 
-  const defaultContractAddress = ConnextContract[sourceChainId]
-  const destinationDomain = chainIdToDomain(destinationChainId)
-  const requiresWrapperMultisend = tokenAddress === ETH_TOKEN_ADDRESS
+  const xcallContractAddress = ConnextContract[sourceChainId]
+  const destinationDomain = destinationChainId
+    ? chainIdToDomain(destinationChainId)
+    : undefined
+  const multiSendContractAddress =
+    getDeployedMultisendContract(sourceChainId)?.address
+  const ethUsedIn = tokenAddress === ETH_TOKEN_ADDRESS
 
-  /* 
-    Connext uses a MultiSend to wrap when briding ETH.
-    https://github.com/connext/monorepo/issues/2905
-    https://github.com/connext/monorepo/issues/4218
-    Contract addresses: https://github.com/search?q=repo%3Aconnext%2Fmonorepo%20MultiSend.json&type=code
-  */
+  if (!xcallContractAddress) {
+    throw new Error(`No xcall contract deployed on chain ${sourceChainId}`)
+  }
 
-  if (requiresWrapperMultisend) {
-    const multiSendContract = getDeployedMultisendContract(sourceChainId)
-
-    if (!multiSendContract) {
-      throw new Error(
-        `No multisend contract deployed on chain ${sourceChainId}`,
-      )
-    }
-
-    const wethAddress = await getWETHAddress(sourceChainId)
-
-    if (!wethAddress) {
-      throw new Error(`No WETH address found on chain ${sourceChainId}`)
-    }
-
-    return compressJson({
-      chainId: sourceChainId,
-      to: multiSendContract.address,
-      value: amount,
-      input: {
-        $abi: MultisendAbi,
-        transactions: {
-          $regex: wethAddress.slice(2),
-        },
-      },
-    })
+  if (!multiSendContractAddress) {
+    throw new Error(`No multisend contract deployed on chain ${sourceChainId}`)
   }
 
   return compressJson({
     chainId: sourceChainId,
-    to: contractAddress || defaultContractAddress,
+    to: {
+      $or: [
+        xcallContractAddress.toLowerCase(),
+        multiSendContractAddress.toLowerCase(),
+      ],
+    },
+    from: recipient,
+    value: ethUsedIn ? amount : undefined,
     input: {
-      $abi: XCALL_ABI_FRAGMENTS,
-      _destination: Number(destinationDomain),
-      _asset: tokenAddress,
-      _amount: amount,
-      _to: recipient,
+      $or: [
+        {
+          $abi: MultisendAbi,
+          transactions: {
+            $regex: recipient?.toLowerCase().slice(2),
+          },
+        },
+        {
+          $abi: XCALL_ABI_FRAGMENTS,
+          _destination: destinationDomain
+            ? Number(destinationDomain)
+            : undefined,
+          _asset: tokenAddress,
+          _amount: amount,
+          _delegate: recipient,
+        },
+      ],
     },
   })
 }