+++ packages/symbiosis/src/utils.ts
@@ -1,19 +1,20 @@
-import type { ActionParams, FilterOperator } from '@rabbitholegg/questdk'
+import type { ActionParams } from '@rabbitholegg/questdk'
 import type { Address, Hash } from 'viem'
 
-export enum Tokens {
-  ARB = '0x912CE59144191C1204E64559FE8253a0e49E6548',
-  DAI = '0xDA10009cBd5D07dd0CeCc66161FC93D7c9000da1',
-  ETH = '0x0000000000000000000000000000000000000000',
-  GMX = '0xfc5A1A6EB076a2C7aD06eD22C90d7E710E35ad0a',
-  GRAIL = '0x3d9907F9a368ad0a51Be60f7Da3b97cf940982D8',
-  MIM = '0xFEa7a6a0B346362BF88A9e4A88416B77a57D6c2A',
-  USDC = '0xaf88d065e77c8cC2239327C5EDb3A432268e5831',
-  USDCE = '0xFF970A61A04b1cA14834A43f5dE4533eBDDB5CC8',
-  USDT = '0xFd086bC7CD5C481DCC9C85ebE478A1C0b69FCbb9',
-  WBTC = '0x2f2a2543B76A4166549F7aaB2e75Bef0aefC5B0f',
-  WETH = '0x82aF49447D8a07e3bd95BD0d56f35241523fBab1',
-  SIZE = '0x939727d85D99d0aC339bF1B76DfE30Ca27C19067',
+export enum Chains {
+  ETHEREUM = 1,
+  OPTIMISM = 10,
+  BINANCE_SMART_CHAIN = 56,
+  GNOSIS = 100,
+  POLYGON_POS = 137,
+  ZK_SYNC_ERA = 324,
+  POLYGON_ZK = 1101,
+  MANTLE = 5000,
+  BASE = 8453,
+  ARBITRUM_ONE = 42161,
+  AVALANCHE = 43114,
+  LINEA = 59144,
+  SCROLL = 534352,
 }
 
 interface Transaction {
@@ -58,22 +59,5 @@ export function createTestCase<T extends ActionParams>(
     transaction: testParams.transaction,
     params: { ...testParams.params, ...overrides },
     description,
-  }
-}
-
-export const buildPathQuery = (tokenIn?: string, tokenOut?: string) => {
-  // v2 paths are formatted as [<token>, <token>]
-  const conditions: FilterOperator[] = []
-
-  if (tokenIn) {
-    conditions.push({ $first: tokenIn })
-  }
-
-  if (tokenOut) {
-    conditions.push({ $last: tokenOut })
-  }
-
-  return {
-    $and: conditions,
   }
 }