+++ packages/symbiosis/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rabbitholegg/questdk-plugin-symbiosis
 
+## 1.0.0-alpha.9
+
+### Minor Changes
+
+- [#150](https://github.com/rabbitholegg/questdk-plugins/pull/150) [`9430b81ef1ddabd7b9f66b9dd978b72a2ceac7ea`](https://github.com/rabbitholegg/questdk-plugins/commit/9430b81ef1ddabd7b9f66b9dd978b72a2ceac7ea) Thanks [@mmackz](https://github.com/mmackz)! - update and pin symbiosis sdk and add support for new metaburn ABI
+
+## 1.0.0-alpha.8
+
+### Patch Changes
+
+- [#147](https://github.com/rabbitholegg/questdk-plugins/pull/147) [`0f2d0d5c4073d6befe2d1a1bd7fa48962010201f`](https://github.com/rabbitholegg/questdk-plugins/commit/0f2d0d5c4073d6befe2d1a1bd7fa48962010201f) Thanks [@mmackz](https://github.com/mmackz)! - pin sdk version
+
 ## 1.0.0-alpha.7
 
 ### Minor Changes